remove interfaces file
