--- src/interfaces.ts
@@ -1,9 +1,0 @@
-export interface ColumnDefinition {
-  udtName: string;
-  nullable: boolean;
-  tsType?: string;
-}
-
-export interface TableDefinition {
-  [columnName: string]: ColumnDefinition;
-}
